{{cookiecutter.project_slug}}/requirements/develop.txt
# Development dependencies
-r _base.txt

django-debug-toolbar

django-test-plus

pytest-django
pytest-sugar
